# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: ShariaComply AI API
  description: |
    ShariaComply AI is an intelligent compliance assistant for Islamic banking professionals.
    It provides RAG (Retrieval-Augmented Generation) powered responses with citations from
    BNM (Bank Negara Malaysia) regulations and AAOIFI Shariah Standards.

    ## Features
    - Chat with AI agent for Islamic banking compliance queries
    - Search BNM regulations and policy documents
    - Search AAOIFI Shariah Standards
    - Compare standards across jurisdictions
    - Visual citations with document highlights

    ## Rate Limiting
    API endpoints are rate-limited to ensure fair usage:
    - Chat: 60 requests per minute
    - Conversations: 120 requests per minute
    - Documents: 100 requests per minute

    Rate limit headers are included in responses:
    - `X-RateLimit-Limit`: Maximum requests allowed
    - `X-RateLimit-Remaining`: Remaining requests in window
    - `X-RateLimit-Reset`: Unix timestamp when limit resets
  version: 1.0.0
  contact:
    name: Alqut Digital
    url: https://github.com/alqutdigital/islamic-banking-agent
  license:
    name: Proprietary
    url: https://alqutdigital.com/license
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.shariacomply.ai
    description: Production server
...
paths:
...
  /api/v1/documents/{id}:
    get:
      tags:
        - Documents
      summary: Get document metadata
      description: Retrieve metadata for a specific document.
      operationId: getDocument
      parameters:
        - name: id
          in: path
          required: true
          description: Document UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          description: Invalid document ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/documents/{id}/download:
    get:
      tags:
        - Documents
      summary: Download a document
      description: |
        Get a download URL for a document or its visual representations.

        Download types:
        - `original`: Original document (PDF/HTML)
        - `page`: Specific page as image
        - `highlighted`: Page with highlighted citation
        - `thumbnail`: Thumbnail image of page
      operationId: downloadDocument
      parameters:
        - name: id
          in: path
          required: true
          description: Document UUID
          schema:
            type: string
            format: uuid
        - name: type
          in: query
          description: Type of download
          schema:
            type: string
            enum:
              - original
              - page
              - highlighted
              - thumbnail
            default: original
        - name: page
          in: query
          description: Page number (required for page/thumbnail types)
          schema:
            type: integer
            minimum: 1
        - name: chunk_id
          in: query
          description: Chunk ID (required for highlighted type)
          schema:
            type: string
            format: uuid
      responses:
        '307':
          description: Temporary redirect to signed download URL
          headers:
            Location:
              description: Signed URL for download
              schema:
                type: string
                format: uri
        '400':
          description: Bad request (missing required parameters)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Document or file not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    Document:
      type: object
      required:
        - id
        - source_type
        - title
        - language
        - is_active
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        source_type:
          type: string
          enum:
            - bnm
            - aaoifi
            - other
        title:
          type: string
        file_name:
          type: string
        file_path:
          type: string
        original_url:
          type: string
          format: uri
        category:
          type: string
        standard_number:
          type: string
        effective_date:
          type: string
          format: date
        total_pages:
          type: integer
        language:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/APIError'
    APIError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Error code
          enum:
            - BAD_REQUEST
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - METHOD_NOT_ALLOWED
            - CONFLICT
            - VALIDATION_ERROR
            - RATE_LIMIT_EXCEEDED
            - INTERNAL_ERROR
            - SERVICE_UNAVAILABLE
        message:
          type: string
          description: Human-readable error message
        details:
          description: Additional error details
          oneOf:
            - type: object
            - type: array
              items:
                $ref: '#/components/schemas/ValidationError'
    ValidationError:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
          description: Field that failed validation
        message:
          type: string
          description: Validation error message